Final polish to Interpose facade
