+++ Sources/InterposeKit/Interpose.swift
@@ -1,27 +1,97 @@
-import Foundation
+import ObjectiveC
 
 /// Interpose is a modern library to swizzle elegantly in Swift.
-///
-/// Methods are hooked via replacing the implementation, instead of the usual exchange.
-/// Supports both swizzling classes and individual objects.
-final public class Interpose {
+public enum Interpose {
     
-    public init() {}
+    // ============================================================================ //
+    // MARK: Class Hooks
+    // ============================================================================ //
     
-}
-
-// MARK: Logging
-
-extension Interpose {
+    public static func prepareHook<MethodSignature, HookSignature>(
+        on `class`: AnyClass,
+        for selector: Selector,
+        methodSignature: MethodSignature.Type,
+        hookSignature: HookSignature.Type,
+        build: @escaping HookBuilder<MethodSignature, HookSignature>
+    ) throws -> Hook {
+        try Hook(
+            target: .class(`class`),
+            selector: selector,
+            build: build
+        )
+    }
+    
+    @discardableResult
+    public static func applyHook<MethodSignature, HookSignature>(
+        on `class`: AnyClass,
+        for selector: Selector,
+        methodSignature: MethodSignature.Type,
+        hookSignature: HookSignature.Type,
+        build: @escaping HookBuilder<MethodSignature, HookSignature>
+    ) throws -> Hook {
+        let hook = try prepareHook(
+            on: `class`,
+            for: selector,
+            methodSignature: methodSignature,
+            hookSignature: hookSignature,
+            build: build
+        )
+        try hook.apply()
+        return hook
+    }
+    
+    // ============================================================================ //
+    // MARK: Object Hooks
+    // ============================================================================ //
+    
+    public static func prepareHook<MethodSignature, HookSignature>(
+        on object: NSObject,
+        for selector: Selector,
+        methodSignature: MethodSignature.Type,
+        hookSignature: HookSignature.Type,
+        build: @escaping HookBuilder<MethodSignature, HookSignature>
+    ) throws -> Hook {
+        try Hook(
+            target: .object(object),
+            selector: selector,
+            build: build
+        )
+    }
+    
+    @discardableResult
+    public static func applyHook<MethodSignature, HookSignature>(
+        on object: NSObject,
+        for selector: Selector,
+        methodSignature: MethodSignature.Type,
+        hookSignature: HookSignature.Type,
+        build: @escaping HookBuilder<MethodSignature, HookSignature>
+    ) throws -> Hook {
+        let hook = try prepareHook(
+            on: object,
+            for: selector,
+            methodSignature: methodSignature,
+            hookSignature: hookSignature,
+            build: build
+        )
+        try hook.apply()
+        return hook
+    }
+    
+    // ============================================================================ //
+    // MARK: Logging
+    // ============================================================================ //
+    
+    /// The flag indicating whether logging is enabled.
     public static var isLoggingEnabled = false
-
-    static func log(_ object: Any) {
-        if isLoggingEnabled {
-            print("[InterposeKit] \(object)")
+    
+    internal static func log(_ message: String) {
+        if self.isLoggingEnabled {
+            print("[InterposeKit] \(message)")
         }
     }
     
-    static func fail(_ message: String) -> Never {
+    internal static func fail(_ message: String) -> Never {
         fatalError("[InterposeKit] \(message)")
     }
+    
 }
