--- Sources/InterposeKit/__Interpose.swift
@@ -1,79 +1,0 @@
-import ObjectiveC
-
-extension Interpose {
-    
-    // ============================================================================ //
-    // MARK: Class Hooks
-    // ============================================================================ //
-    
-    public static func prepareHook<MethodSignature, HookSignature>(
-        on `class`: AnyClass,
-        for selector: Selector,
-        methodSignature: MethodSignature.Type,
-        hookSignature: HookSignature.Type,
-        build: @escaping HookBuilder<MethodSignature, HookSignature>
-    ) throws -> Hook {
-        try Hook(
-            target: .class(`class`),
-            selector: selector,
-            build: build
-        )
-    }
-    
-    @discardableResult
-    public static func applyHook<MethodSignature, HookSignature>(
-        on `class`: AnyClass,
-        for selector: Selector,
-        methodSignature: MethodSignature.Type,
-        hookSignature: HookSignature.Type,
-        build: @escaping HookBuilder<MethodSignature, HookSignature>
-    ) throws -> Hook {
-        let hook = try prepareHook(
-            on: `class`,
-            for: selector,
-            methodSignature: methodSignature,
-            hookSignature: hookSignature,
-            build: build
-        )
-        try hook.apply()
-        return hook
-    }
-    
-    // ============================================================================ //
-    // MARK: Object Hooks
-    // ============================================================================ //
-    
-    public static func prepareHook<MethodSignature, HookSignature>(
-        on object: NSObject,
-        for selector: Selector,
-        methodSignature: MethodSignature.Type,
-        hookSignature: HookSignature.Type,
-        build: @escaping HookBuilder<MethodSignature, HookSignature>
-    ) throws -> Hook {
-        try Hook(
-            target: .object(object),
-            selector: selector,
-            build: build
-        )
-    }
-    
-    @discardableResult
-    public static func applyHook<MethodSignature, HookSignature>(
-        on object: NSObject,
-        for selector: Selector,
-        methodSignature: MethodSignature.Type,
-        hookSignature: HookSignature.Type,
-        build: @escaping HookBuilder<MethodSignature, HookSignature>
-    ) throws -> Hook {
-        let hook = try prepareHook(
-            on: object,
-            for: selector,
-            methodSignature: methodSignature,
-            hookSignature: hookSignature,
-            build: build
-        )
-        try hook.apply()
-        return hook
-    }
-    
-}
